refactored closeRelease and removed !important inline styling from closeRelease in hubs
